test: add tests to ensure no future deadlocks may occur
