+++ internal/graph/check_test.go
@@ -60,237 +60,6 @@ var (
 	}
 )
 
-func TestAvoidSingleflightDeadlockWithinRequest(t *testing.T) {
-	ds := memory.New()
-	t.Cleanup(ds.Close)
-
-	storeID := ulid.Make().String()
-
-	err := ds.Write(context.Background(), storeID, nil, []*openfgav1.TupleKey{
-		tuple.NewTupleKey("document:1", "editor1", "document:1#viewer1"),
-	})
-	require.NoError(t, err)
-
-	model := testutils.MustTransformDSLToProtoWithID(`model
-			schema 1.1
-
-		  type user
-
-		  type document
-			relations
-			  define editor1: [user, document#viewer1]
-
-			  define viewer2: [document#viewer1] or editor1
-			  define viewer1: [user] or viewer2
-			  define can_view: viewer1 or editor1`)
-
-	err = ds.WriteAuthorizationModel(context.Background(), storeID, model)
-	require.NoError(t, err)
-
-	ctx := typesystem.ContextWithTypesystem(
-		context.Background(),
-		typesystem.New(model),
-	)
-	ctx = storage.ContextWithRelationshipTupleReader(ctx, ds)
-
-	cycleCheckResolver := NewCycleDetectionCheckResolver()
-	localCheckResolver := NewLocalChecker()
-
-	cycleCheckResolver.SetDelegate(localCheckResolver)
-	localCheckResolver.SetDelegate(cycleCheckResolver)
-
-	t.Cleanup(func() {
-		goleak.VerifyNone(t)
-		cycleCheckResolver.Close()
-		localCheckResolver.Close()
-	})
-
-	resp, err := cycleCheckResolver.ResolveCheck(ctx, &ResolveCheckRequest{
-		StoreID:              storeID,
-		AuthorizationModelID: model.GetId(),
-		TupleKey:             tuple.NewTupleKey("document:1", "can_view", "user:jon"),
-		ResolutionMetadata: &ResolutionMetadata{
-			Depth: 25,
-		},
-	})
-	require.NoError(t, err)
-	require.False(t, resp.GetAllowed())
-}
-
-func TestAvoidSingleflightDeadlockAcrossRequests(t *testing.T) {
-	ds := memory.New()
-	t.Cleanup(ds.Close)
-
-	storeID := ulid.Make().String()
-
-	err := ds.Write(context.Background(), storeID, nil, []*openfgav1.TupleKey{
-		tuple.NewTupleKey("document:1", "viewer", "document:1#viewer"),
-		tuple.NewTupleKey("document:1", "editor", "document:1#viewer"),
-		tuple.NewTupleKey("document:1", "editor", "user:andres"),
-	})
-	require.NoError(t, err)
-
-	model := testutils.MustTransformDSLToProtoWithID(`model
-                       schema 1.1
-
-                       type user
-
-                       type document
-                         relations
-                               define viewer: [user, document#viewer] or editor
-                               define editor: [user, document#viewer]`)
-
-	err = ds.WriteAuthorizationModel(context.Background(), storeID, model)
-	require.NoError(t, err)
-
-	ctx := typesystem.ContextWithTypesystem(
-		context.Background(),
-		typesystem.New(model),
-	)
-	ctx = storage.ContextWithRelationshipTupleReader(ctx, ds)
-
-	cycleCheckResolver := NewCycleDetectionCheckResolver()
-	localCheckResolver := NewLocalChecker()
-
-	cycleCheckResolver.SetDelegate(localCheckResolver)
-	localCheckResolver.SetDelegate(cycleCheckResolver)
-
-	t.Cleanup(func() {
-		goleak.VerifyNone(t)
-		cycleCheckResolver.Close()
-		localCheckResolver.Close()
-	})
-
-	var wg sync.WaitGroup
-
-	wg.Add(2)
-
-	var resp1 *ResolveCheckResponse
-	var err1 error
-	go func() {
-		defer wg.Done()
-
-		resp1, err1 = cycleCheckResolver.ResolveCheck(ctx, &ResolveCheckRequest{
-			StoreID:              storeID,
-			AuthorizationModelID: model.GetId(),
-			TupleKey:             tuple.NewTupleKey("document:1", "editor", "user:jon"),
-			ResolutionMetadata: &ResolutionMetadata{
-				Depth: 25,
-			},
-		})
-	}()
-
-	var resp2 *ResolveCheckResponse
-	var err2 error
-	go func() {
-		defer wg.Done()
-
-		resp2, err2 = cycleCheckResolver.ResolveCheck(ctx, &ResolveCheckRequest{
-			StoreID:              storeID,
-			AuthorizationModelID: model.GetId(),
-			TupleKey:             tuple.NewTupleKey("document:1", "viewer", "user:jon"),
-			ResolutionMetadata: &ResolutionMetadata{
-				Depth: 25,
-			},
-		})
-	}()
-
-	wg.Wait()
-
-	require.NoError(t, err1)
-	require.False(t, resp1.GetAllowed())
-
-	require.NoError(t, err2)
-	require.False(t, resp2.GetAllowed())
-}
-
-func TestThreeProngThroughVariousLayers(t *testing.T) {
-	ds := memory.New()
-	t.Cleanup(ds.Close)
-
-	for i := 0; i < 100; i++ {
-		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
-			storeID := ulid.Make().String()
-
-			err := ds.Write(context.Background(), storeID, nil, []*openfgav1.TupleKey{
-				tuple.NewTupleKey("module:a", "owner", "user:anne"),
-				tuple.NewTupleKey("folder:a", "parent", "module:a"),
-				tuple.NewTupleKey("document:a", "parent", "folder:a"),
-				tuple.NewTupleKey("module:b", "parent", "document:a"),
-				tuple.NewTupleKey("folder:b", "parent", "module:b"),
-				tuple.NewTupleKey("document:b", "parent", "folder:b"),
-				tuple.NewTupleKey("module:a", "parent", "document:b"),
-			})
-			require.NoError(t, err)
-
-			model := testutils.MustTransformDSLToProtoWithID(`model
-			schema 1.1
-
-		  type user
-		  type module
-		  relations
-			  define owner: [user] or owner from parent
-			  define parent: [document, module]
-			  define viewer: [user] or owner or viewer from parent
-		  type folder
-		  relations
-			  define owner: [user] or owner from parent
-			  define parent: [module, folder]
-			  define viewer: [user] or owner or viewer from parent
-		  type document
-		  relations
-			  define owner: [user] or owner from parent
-			  define parent: [folder, document]
-			  define viewer: [user] or owner or viewer from parent`)
-
-			err = ds.WriteAuthorizationModel(context.Background(), storeID, model)
-			require.NoError(t, err)
-
-			fmt.Printf("(store, model), (%s, %s)\n", storeID, model.GetId())
-
-			ctx := typesystem.ContextWithTypesystem(
-				context.Background(),
-				typesystem.New(model),
-			)
-			ctx = storage.ContextWithRelationshipTupleReader(ctx, ds)
-
-			cycleCheckResolver := NewCycleDetectionCheckResolver()
-			localCheckResolver := NewLocalChecker()
-
-			cycleCheckResolver.SetDelegate(localCheckResolver)
-			localCheckResolver.SetDelegate(cycleCheckResolver)
-
-			t.Cleanup(func() {
-				goleak.VerifyNone(t)
-				cycleCheckResolver.Close()
-				localCheckResolver.Close()
-			})
-
-			tupleKeys := []*openfgav1.TupleKey{
-				tuple.NewTupleKey("module:a", "viewer", "user:anne"),
-				tuple.NewTupleKey("module:b", "viewer", "user:anne"),
-				tuple.NewTupleKey("folder:a", "viewer", "user:anne"),
-				tuple.NewTupleKey("folder:b", "viewer", "user:anne"),
-				tuple.NewTupleKey("document:a", "viewer", "user:anne"),
-				tuple.NewTupleKey("document:b", "viewer", "user:anne"),
-			}
-
-			for _, tupleKey := range tupleKeys {
-				resp, err := cycleCheckResolver.ResolveCheck(ctx, &ResolveCheckRequest{
-					StoreID:              storeID,
-					AuthorizationModelID: model.GetId(),
-					TupleKey:             tupleKey,
-					ResolutionMetadata: &ResolutionMetadata{
-						Depth: 25,
-					},
-				})
-				require.NoError(t, err)
-				require.True(t, resp.GetAllowed())
-			}
-		})
-	}
-}
-
 func TestExclusionCheckFuncReducer(t *testing.T) {
 	t.Cleanup(func() {
 		goleak.VerifyNone(t)
